75-FEAT: users can now edit customise their settings
