--- apps/meally/src/app/(user)/[profile]/settings/account/page.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function AccountPage() {
-  return <div>Coming soon....</div>;
-}
